Recompute Content-Length on compression
Content-Length holds the number of bytes in an HTTP response after
encoding and compression. Our compress_response() function breaks that:
when the response it's compressing already includes a Content-Length
header, it may shorten the body but leave the header unchanged.

I noticed this because of #465, since the textual error message proxied
from Reddit gets compressed but already has Content-Length. The issue is
masked in normal operation because we don't compress the binary image
data that Reddit is supposed to return. In debug builds, hyper catches
the issue[1] and panics, but in release builds we return a nonconformant
HTTP response.

Fix the issue by removing the header if we choose to compress, letting
hyper compute the new value for us.

[1] https://github.com/hyperium/hyper/blob/v0.14.32/src/proto/h1/role.rs#L708-L725